requirements.txt
feedparser==6.0.11
httpx==0.27.0
gspread==6.1.2
google-auth==2.29.0
openai>=1.0.0
python-dotenv==1.0.1
python-telegram-bot==21.3
